Toastがモーダルを閉じた際に一緒に消えないように修正
